move argwhere to transform
